pyproject.toml
[tool.poetry]
name = "cad"
version = "0.1.0"
description = ""
authors = ["Daniel Winkler <dan@danwink.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10,<3.13"
solidpython = "^1.1.3"
numba = "^0.57.1"
PyMCubes = "^0.1.4"
scipy = "^1.11.1"
numpy-stl = "^3.0.1"
tqdm = "^4.66.1"
pyclipper = "^1.3.0.post4"
shapely = "^2.0.1"
noise = "^1.2.2"
networkx = "^3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
